Update to the schedule implementation
Season contains the start and end dates; the time intervals at which an activity is given; and days in the week at which the activities occur. A season can be associated with multiple activities or multiple instances of an activity that occur at the same location in different times.
Schedule has its own start time and end time. This represents the time open and closing hours.